requirements.txt
aiohttp==3.13.2
beautifulsoup4==4.14.3
pandas==2.3.3
python-dotenv==1.2.1
requests==2.32.5